XRDDEV-1449 work on refactoring
