+++ src/proxy-ui-api/frontend/tests/e2e/pages/tabs/keysTab.js
@@ -1,0 +1,78 @@
+/*
+ * The MIT License
+ * Copyright (c) 2019- Nordic Institute for Interoperability Solutions (NIIS)
+ * Copyright (c) 2018 Estonian Information System Authority (RIA),
+ * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
+ * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in
+ * all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+ * THE SOFTWARE.
+ */
+
+var keysTabCommands = {
+  openSignAndAuthKeys: function () {
+    this.click('@signAndAuthKeysTab');
+    return this;
+  },
+  openAPIKeys: function () {
+    this.click('@APIKeysTab');
+    return this;
+  },
+  openSecurityServerTLSKey: function () {
+    this.click('@securityServerTLSKeyTab');
+    return this;
+  },
+};
+
+const keysTab = {
+  url: process.env.VUE_DEV_SERVER_URL,
+  selector: '//div[.//a[contains(@class, "v-tab--active") and contains(text(), "Keys and certificates")]]//div[contains(@class, "base-full-width")]',
+  commands: keysTabCommands,
+  elements: {
+    signAndAuthKeysTab: {
+      selector: '//div[contains(@class, "v-tabs-bar__content")]//a[text()=" SIGN and AUTH Keys "]',
+      locateStrategy: 'xpath',
+    },
+    APIKeysTab: {
+      selector: '//div[contains(@class, "v-tabs-bar__content")]//a[text()=" API Keys "]',
+      locateStrategy: 'xpath',
+    },
+    securityServerTLSKeyTab: {
+      selector: '//div[contains(@class, "v-tabs-bar__content")]//a[text()=" Security Server TLS Key "]',
+      locateStrategy: 'xpath',
+    },
+    tokenName: {
+      selector: '//*[contains(@data-test, "token-name")]',
+      locateStrategy: 'xpath',
+    },
+    createAPIKeyButton: {
+      selector: '//*[contains(@data-test, "api-key-create-key-button")]',
+      locateStrategy: 'xpath',
+    },
+    generateKeyButton: {
+      selector: '//*[contains(@data-test, "security-server-tls-certificate-generate-key-button")]',
+      locateStrategy: 'xpath',
+    },
+    exportCertButton: {
+      selector: '//*[contains(@data-test, "security-server-tls-certificate-export-certificate-button")]',
+      locateStrategy: 'xpath',
+    },
+  },
+};
+
+module.exports = keysTab;
